reorganize nav components
